warnings
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@7882 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Network/DevPCNet.cpp
@@ -1822,7 +1822,7 @@ static void pcnetReceiveNoSync(PCNetState *pData, const uint8_t *buf, size_t siz
             /*if (!CSR_LAPPEN(pData))*/
                 rmd.rmd1.stp = 1;
 
-            int count = RT_MIN(4096 - (int)rmd.rmd1.bcnt, size);
+            size_t count = RT_MIN(4096 - (size_t)rmd.rmd1.bcnt, size);
             RTGCPHYS32 rbadr = PHYSADDR(pData, rmd.rmd0.rbadr);
 #if 0
             if (pData->fPrivIfEnabled)
@@ -1861,7 +1861,7 @@ static void pcnetReceiveNoSync(PCNetState *pData, const uint8_t *buf, size_t siz
                 crda = next_crda;
                 rmd  = next_rmd;
 
-                count = RT_MIN(4096 - (int)rmd.rmd1.bcnt, size);
+                count = RT_MIN(4096 - (size_t)rmd.rmd1.bcnt, size);
                 RTGCPHYS32 rbadr = PHYSADDR(pData, rmd.rmd0.rbadr);
 #if 0
                 if (pData->fPrivIfEnabled)
